requirements.txt
alchemy==20.5
alembic==1.16.1
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.6.15
chardet==3.0.4
click==8.2.1
dnspython==2.7.0
docutils==0.21.2
email_validator==2.2.0
fastapi==0.115.12
filelock==3.0.12
h11==0.16.0
idna==2.8
lockfile==0.12.2
Mako==1.3.10
MarkupSafe==3.0.2
pydantic==2.11.7
pydantic_core==2.33.2
python-daemon==2.2.4
python-dotenv==1.1.0
requests==2.22.0
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.46.2
typing-inspection==0.4.1
typing_extensions==4.14.0
urllib3==1.25.11
uvicorn==0.34.3
